refactor checkdaemon ipc and readExesVersion
